Desktop/legion-website/requirements.txt
# Основные зависимости для CrewAI 0.159.0
crewai==0.159.0

# LangChain и связанные пакеты
langchain==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-text-splitters==0.3.9

# OpenAI и LLM - используем более старые совместимые версии
openai==1.68.2
litellm==1.74.9

# Утилиты
python-dotenv==1.1.1

# Дополнительные зависимости CrewAI
chromadb>=0.5.23
instructor>=1.3.3
json-repair==0.25.2
json5>=0.10.0
jsonref>=1.1.0
onnxruntime==1.22.0
openpyxl>=3.1.5
opentelemetry-api>=1.30.0
opentelemetry-exporter-otlp-proto-http>=1.30.0
opentelemetry-sdk>=1.30.0
pdfplumber>=0.11.4
portalocker==2.7.0
pydantic>=2.4.2
pyjwt>=2.9.0
pyvis>=0.3.2
regex>=2024.9.11
tokenizers>=0.20.3
tomli-w>=1.1.0
tomli>=2.0.2
uv>=0.4.25

# Зависимости для работы с вебом
aiohttp>=3.10
httpx>=0.23.0
requests>=2.7
